# backend/cube/docs/swagger.yaml
basePath: /api
definitions:
...
  WeightStatInfoResponse:
    properties:
      edit:
        items:
          $ref: '#/definitions/weightsstat.StatInfo'
        type: array
      load:
        items:
          $ref: '#/definitions/weightsstat.StatInfo'
        type: array
    type: object
...
  weightsstat.StatInfo:
    properties:
      id:
        example: 3f8bf2a3-01cc-4c4a-9759-86cec9cf8da9
        type: string
      time:
        example: 2006-01-02T15:04:05Z07:00
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: This is cube server.
  license:
    name: MIT
    url: https://mit-license.org/
  title: Cube API
  version: "1.0"
paths:
...
  /v1/weights/stats:
    get:
      description: Get such user stat info as registration and edit stat per period
      parameters:
      - description: Time to start from, RFC3339
        format: '"2006-01-02T15:04:05Z07:00"'
        in: query
        name: from
        type: string
      - description: Time to stop at, RFC3339
        format: '"2006-01-02T15:04:05Z07:00"'
        in: query
        name: to
        type: string
      - description: Search for load stat
        in: query
        name: load
        type: boolean
      - description: Search for update stats
        in: query
        name: update
        type: boolean
      responses:
        "200":
          description: Weights stat info found
          schema:
            items:
              $ref: '#/definitions/WeightStatInfoResponse'
            type: array
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "500":
          description: Failed to get user stat info
      security:
      - ApiKeyAuth: []
      summary: Get users stat info
      tags:
      - weights
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"